requirements.txt
absl-py==2.1.0
addict==2.4.0
basicsr==1.4.2
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
facexlib==0.3.0
filelock==3.16.1
filterpy==1.4.5
fonttools==4.55.1
fsspec==2024.10.0
future==1.0.0
gfpgan==1.3.8
grpcio==1.68.1
idna==3.10
imageio==2.36.1
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.5.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.3
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.29.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
realesrgan==0.3.0
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
six==1.17.0
sympy==1.13.3
tb-nightly==2.19.0a20241121
tensorboard-data-server==0.7.2
tifffile==2024.9.20
tomli==2.2.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
yapf==0.43.0
